requirements.txt
# Rick Case Honda AI Agent - Python Dependencies

# Telegram Bot
python-telegram-bot==20.7

# Environment Variables
python-dotenv==1.0.0

# Data Processing
pandas==2.1.4

# AI/ML Libraries
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
langchain-core==0.1.10

# Vector Database
pinecone-client==3.0.0

# OpenAI
openai==1.7.0

# PDF Processing
pypdf==3.17.4
